cleanup: move WORKSPACE_TYPE constant to core/utils/constants (#129)
* cleanup: move WORKSPACE_TYPE constant to core/utils/constants


* fix: bootstrap


---------
