+++ src/core/public/index.ts
@@ -356,6 +356,11 @@ export {
   WorkspaceObservables,
 } from './workspace';
 
-export { WorkspacePermissionMode, PUBLIC_WORKSPACE_ID, MANAGEMENT_WORKSPACE_ID } from '../utils';
+export {
+  WorkspacePermissionMode,
+  PUBLIC_WORKSPACE_ID,
+  MANAGEMENT_WORKSPACE_ID,
+  WORKSPACE_TYPE,
+} from '../utils';
 
-export { getWorkspaceIdFromUrl, WORKSPACE_TYPE } from './utils';
+export { getWorkspaceIdFromUrl } from './utils';
