+++ src/core/server/index.ts
@@ -323,7 +323,6 @@ export {
   SavedObjectsShareObjects,
   SavedObjectsAddToWorkspacesOptions,
   SavedObjectsAddToWorkspacesResponse,
-  WORKSPACE_TYPE,
   Permissions,
   ACL,
   SavedObjectsPermissionControlContract,
@@ -359,6 +358,7 @@ export {
   WorkspacePermissionMode,
   PUBLIC_WORKSPACE_ID,
   MANAGEMENT_WORKSPACE_ID,
+  WORKSPACE_TYPE,
 } from '../utils';
 
 export {
