--- src/core/server/saved_objects/constants.ts
@@ -1,6 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export const WORKSPACE_TYPE = 'workspace';
