+++ src/core/utils/constants.ts
@@ -3,6 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+export const WORKSPACE_TYPE = 'workspace';
+
 export const WORKSPACE_PATH_PREFIX = '/w';
 
 export enum WorkspacePermissionMode {
